+++ src/Pages/Process/FillData.jsx
@@ -68,7 +68,7 @@ const FillData=()=>{
             setServices(response.data.ContentData);
         }catch({err , response}){
             setLoading(false);
-            if(response.status===401){
+            if(response && response.status===401){
                 localStorage.clear();
                 history.push("/login");
                 toast.error("شما از برنامه خارج شده اید",{
