+++ src/Pages/Process/SelectOrg.jsx
@@ -53,7 +53,7 @@ const SelectOrg=()=>{
             console.log(response.data.ContentData);
             setBrands(response.data.ContentData);
         }catch({err , response}){
-            if(response.status===401){
+            if(response && response.status===401){
                 localStorage.clear();
                 history.push("/login");
                 toast.error("شما از برنامه خارج شده اید",{
